move `common/src/estimate_size` to `common/estimate_size`
