+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
   "src/cmd_all",
   "src/common",
   "src/common/common_service",
+  "src/common/estimate_size",
   "src/common/fields-derive",
   "src/common/heap_profiling",
   "src/common/metrics",
@@ -174,6 +175,7 @@ risingwave_batch = { path = "./src/batch" }
 risingwave_cmd = { path = "./src/cmd" }
 risingwave_common = { path = "./src/common" }
 risingwave_common_service = { path = "./src/common/common_service" }
+risingwave_common_estimate_size = { path = "./src/common/estimate_size" }
 risingwave_common_heap_profiling = { path = "./src/common/heap_profiling" }
 risingwave_common_metrics = { path = "./src/common/metrics" }
 risingwave_compactor = { path = "./src/storage/compactor" }
