+++ src/common/estimate_size/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "risingwave_common_estimate_size"
+version = { workspace = true }
+edition = { workspace = true }
+homepage = { workspace = true }
+keywords = { workspace = true }
+license = { workspace = true }
+repository = { workspace = true }
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[package.metadata.cargo-machete]
+ignored = ["workspace-hack"]
+
+[package.metadata.cargo-udeps.ignore]
+normal = ["workspace-hack"]
+
+[dependencies]
+bytes = "1"
+educe = "0.5"
+fixedbitset = "0.5"
+lru = { workspace = true }
+rust_decimal = "1"
+serde_json = "1"
+
+[lints]
+workspace = true
