+++ src/common/estimate_size/src/collections/mod.rs
@@ -23,6 +23,8 @@ pub mod hashmap;
 pub use hashmap::EstimatedHashMap;
 pub mod btreemap;
 pub use btreemap::EstimatedBTreeMap;
+pub mod vec;
+pub use vec::VecWithKvSize;
 
 mod private {
     use super::*;
